Completed Assignment 1
Exported animations from Mixamo, originally broken ones but fixed later on.
Had difficulty with understanding how to make the animations function in the game so I moved onto the 2nd task.

Wanted to make the objects fade back into the game using the material's alpha.

References used for object's material alpha:
https://discussions.unity.com/t/how-to-change-materials-alpha-of-a-gameobject/187476
https://www.reddit.com/r/Unity3D/comments/de2if3/fading_3d_object_in_and_out/
https://docs.unity3d.com/ScriptReference/Material-shader.html

References used for coroutines and timing for object respawning:
https://docs.unity3d.com/Manual/Coroutines.html
https://discussions.unity.com/t/c-how-can-i-let-something-happen-after-a-small-delay/117594

Finally implemented the animations, with mild help from Ryan and a lot of trial and error.
It was a lot of looking at the original code and trying to figure things out from there.

Got camera help from Ryan and Owen.
Camera references:
https://www.youtube.com/watch?v=Cfcr7FtbYQo&t=297s
https://docs.unity3d.com/ScriptReference/LayerMask.html